Yet another command overhaul!
Found a cleaner way to find if a member has the right perms.
Added a send command to send others points.
Groundwork for shop and gifting.
Bot now create files when added to a server.
Changed most instances of users to members.
Allowed server owners to run admin commands regardless of role or being an exception.
Removed looping through all members to see which member ran the command, now just runs it through the author of the message.